2025-12-05_super_taller_cv/requirements.txt
# Core Deep Learning Framework
tensorflow>=2.10.0
keras>=2.10.0
torch>=1.12.0  # Optional: for PyTorch models

# Scientific Computing
numpy>=1.21.0
scipy>=1.7.0
scikit-learn>=1.0.0
pandas>=1.3.0

# Visualization
matplotlib>=3.5.0
seaborn>=0.11.0
plotly>=5.0.0
imageio>=2.9.0
Pillow>=8.0.0

# Web Dashboard
dash>=2.0.0
dash-bootstrap-components>=1.0.0

# Image Processing
opencv-python>=4.5.0
opencv-contrib-python>=4.5.0

# Utilities
tqdm>=4.60.0
pyyaml>=5.4.0
python-dotenv>=0.19.0

# Optional: Enhanced GPU Support
# cudatoolkit>=11.0  # For NVIDIA GPU support
# cudnn>=8.0  # CUDA Deep Neural Network library

# Development
jupyterlab>=3.0.0
ipykernel>=6.0.0

# Testing & Documentation
pytest>=6.2.0
sphinx>=4.0.0
